+++ scripts/build_deps.sh
@@ -6,5 +6,5 @@ cd "$SCRIPT_DIR"
 
 ./build.uv.sh
 ./build.hwloc.sh
-./build.openssl3.sh
+./build.openssl4.sh
 ./build.zlib.sh