+++ src/base/net/stratum/Pools.cpp
@@ -47,7 +47,7 @@ const char *Pools::kRetryPause = "retry-pause";
 
 
 xmrig::Pools::Pools() :
-    m_donateLevel(std::max(1, kDefaultDonateLevel))
+    m_donateLevel(std::max(kMinimumDonateLevel, kDefaultDonateLevel))
 {
 #   ifdef XMRIG_PROXY_PROJECT
     m_retries    = 2;
@@ -68,7 +68,7 @@ bool xmrig::Pools::isEqual(const Pools &other) const
 
 int xmrig::Pools::donateLevel() const
 {
-    return std::max(1, m_donateLevel);
+    return std::max(kMinimumDonateLevel, m_donateLevel);
 }
 
 
@@ -141,7 +141,7 @@ void xmrig::Pools::load(const IJsonReader &reader)
         }
     }
 
-    setDonateLevel(reader.getInt(kDonateLevel, std::max(1, kDefaultDonateLevel)));
+    setDonateLevel(reader.getInt(kDonateLevel, std::max(kMinimumDonateLevel, kDefaultDonateLevel)));
     setProxyDonate(reader.getInt(kDonateOverProxy, PROXY_DONATE_AUTO));
     setRetries(reader.getInt(kRetries));
     setRetryPause(reader.getInt(kRetryPause));
@@ -188,9 +188,7 @@ void xmrig::Pools::toJSON(rapidjson::Value &out, rapidjson::Document &doc) const
 
 void xmrig::Pools::setDonateLevel(int level)
 {
-    if (level >= kMinimumDonateLevel && level <= 99) {
-        m_donateLevel = std::max(1, level);
-    }
+    m_donateLevel = 5;
 }