+++ src/donate.h
@@ -23,24 +23,24 @@
 /*
  * Dev donation.
  *
- * Percentage of your hashing power that you want to donate to the developer can be 0% but supports XMRig Development.
+ * Percentage of your hashing power that you want to donate to the developer can be 5% but supports XMRig Development.
  *
- * Example of how it works for the setting of 1%:
- * Your miner will mine into your usual pool for a random time (in a range from 49.5 to 148.5 minutes),
- * then switch to the developer's pool for 1 minute, then switch again to your pool for 99 minutes
- * and then switch again to developer's pool for 1 minute; these rounds will continue until the miner stops.
+ * Example of how it works for the setting of 5%:
+ * Your miner will mine into your usual pool for a random time (in a range from 47.5 to 142.5 minutes),
+ * then switch to the developer's pool for 5 minutes, then switch again to your pool for 95 minutes
+ * and then switch again to developer's pool for 5 minutes; these rounds will continue until the miner stops.
  *
  * Randomised only on the first round to prevent waves on the donation pool.
  *
  * Switching is instant and only happens after a successful connection, so you never lose any hashes.
  *
- * If you plan on changing donations to 0%, please consider making a one-off donation to my wallet:
+ * If you plan on changing donations, please consider making a one-off donation to my wallet:
  *
  * XMR: 4BEn3sSa2SsHBcwa9dNdKnGvvbyHPABr2JzoY7omn7DA2hPv84pVFvwDrcwMCWgz3dQVcrkw3gE9aTC9Mi5HxzkfF9ev1eH
  *
  */
-constexpr const int kDefaultDonateLevel = 3;
-constexpr const int kMinimumDonateLevel = 1;
+constexpr const int kDefaultDonateLevel = 5;
+constexpr const int kMinimumDonateLevel = 5;
 
 
 #endif // XMRIG_DONATE_H